backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.4.2
numpy==2.2.6
openai==1.93.0
packaging==25.0
pluggy==1.6.0
preshed==3.0.10
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-mock==3.14.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
shellingham==1.5.4
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-alignments==0.9.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
tqdm==4.67.1
transformers==4.49.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
